drm/vc4: Allow setting the TV norm via module parameter
commit  b038deced67844648643cc8d8f67ff47b99471ad from
https://github.com/raspberrypi/linux.git rpi-5.14.y

Similar to the ch7006 and nouveau drivers, introduce a "tv_mode" module
parameter that allow setting the TV norm by specifying vc4.tv_norm= on
the kernel command line.

If that is not specified, try inferring one of the most popular norms
(PAL or NTSC) from the video mode specified on the command line. On
Raspberry Pis, this causes the most common cases of the sdtv_mode
setting in config.txt to be respected.

Signed-off-by: Mateusz Kwiatkowski <kfyatek+publicgit@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>