cpu/amd: inhibit SMP check for qemux86
When booting with kvm enabled on a AMD host, the following
trace is thrown:

  [    0.084519] ------------[ cut here ]------------
  [    0.084519] WARNING: This combination of AMD processors is not suitable for SMP.
  [    0.084519] WARNING: CPU: 1 PID: 0 at /arch/x86/kernel/cpu/amd.c:341 init_amd+0xaee/0xbcc
  [    0.084519] Modules linked in:
  [    0.084519] CPU: 1 PID: 0 Comm: swapper/1 Not tainted 6.6.32-yocto-standard #1
  [    0.084519] Hardware name: QEMU Standard PC (Q35 + ICH9, 2009), BIOS rel-1.16.3-0-ga6ed6b701f0a-prebuilt.qemu.org 04/01/2014

This warning is not valid in our configuration and is unnecesarily
causing issue with debug.

This has been know for some time (10+ years), but no acceptable
solutioon has been found upstream:

   https://lists.gnu.org/archive/html/qemu-devel/2010-03/msg01428.html
   https://lkml.org/lkml/2010/3/30/397

We have a configuration CONFIG_QEMUX86 that has been added for
situations like this. When that value is defined, we inhibit the
warning, but leave it as-is for other BSPs.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>