fs/yaffs2: fix a build warning in yaffs_yaffs2.c when running make allyesconfig
When building linux yocto with allyesconfig, a build warning is
found:

fs/yaffs2/yaffs_yaffs2.c:557:21: error: no previous prototype for
‘yaffs2_do_endian_tnode_copy’ [-Werror=missing-prototypes]
  557 | struct yaffs_tnode *yaffs2_do_endian_tnode_copy(struct yaffs_dev
      *dev,
      |                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~
cc1: all warnings being treated as errors

Inspect the code that the function yaffs2_do_endian_tnode_copy() is
only used in the current file. So add the static modifier to it.

Signed-off-by: Bin Lan <bin.lan.cn@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>